cleanup ProgressReporter

+++ backend/src/main/java/com/bakdata/conquery/util/progressreporter/ProgressReporter.java
@@ -2,13 +2,16 @@ package com.bakdata.conquery.util.progressreporter;
 
 import com.fasterxml.jackson.annotation.JsonValue;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
+
 @JsonDeserialize(as = ImmutableProgressReporter.class)
 public interface ProgressReporter {
+
 	public static ProgressReporter createStarted() {
 		ProgressReporter pr = new ProgressReporterImpl();
 		pr.start();
 		return pr;
 	}
+
 	public static ProgressReporter createWaiting() {
 		return new ProgressReporterImpl();
 	}
@@ -23,14 +26,18 @@ public interface ProgressReporter {
 	long getWaitedSeconds();
 	long getStartTime();
 	boolean isStarted();
-	
-	
+
+
 	double getProgress();
-	ProgressReporter subJob(double steps);
+
+	ProgressReporter subJob(int steps);
 	String getEstimate();
-	void report(double steps);
-	void setMax(double max);
-	double getMax();
+
+	void report(int steps);
+
+	void setMax(int max);
+	int getMax();
+
 	void done();
 	boolean isDone();
 }
